--- aiter/ops/flydsl/kernels/fmha_gfx950/flash_attn_utils_primitives.py
@@ -1,425 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Copyright (c) 2025 FlyDSL Project Contributors
-# Modifications Copyright (C) 2026 Advanced Micro Devices, Inc.
-
-"""Low-level ROCDL / MLIR primitives and tile constants.
-
-Part of the gfx950 dual-wave fp8 (e4m3fn) flash-attention kernel, migrated from
-FlyDSL ``kernels/attention/flash_attn_utils.py`` and restricted to the symbols
-the fp8 path reaches. The bf16/f16 dual-wave, the gfx942 generic path, paged KV,
-and the bias/ALiBi helpers are not part of it and were left behind.
-"""
-
-import math as host_math
-
-import flydsl.compiler as flyc
-import flydsl.expr as fx
-from flydsl._mlir import ir
-from flydsl._mlir.dialects import fly, llvm, vector
-from flydsl.expr import const_expr, range_constexpr, rocdl
-from flydsl.expr.typing import T
-from flydsl.expr.typing import Vector as Vec
-from flydsl.expr.utils.arith import _to_raw as as_mlir_value
-
-from aiter.ops.flydsl.kernels import buffer_ops
-
-_LOG2E = host_math.log2(host_math.e)
-
-
-# gfx950 (MI350/MI355X): 8 XCDs, each with a private ~4 MB L2.
-NUM_XCD_GFX950 = 8
-
-
-LDS_BYTES_GFX950 = 160 * 1024
-
-
-MIN_Q_BLOCKS_XCD_SWIZZLE = 64
-
-
-# The dual-wave 8-wave CTA fixes the q-block height; callers need it to count
-# q-blocks before any traits object exists.
-DUALWAVE_SWP_BLOCK_M = 256
-
-
-def _waitcnt_vm_n(n):
-    """Emit s_waitcnt vmcnt(n) only."""
-    rocdl.s_waitcnt(vmcnt=n)
-
-
-def _s_setprio(val):
-    rocdl.s_setprio(val)
-
-
-def _read_exec_i64():
-    """Read the current wave exec mask, matching Clang's builtin lowering."""
-    true_i1 = fx.Boolean(True).ir_value()
-    return rocdl.ballot(T.i64, true_i1)
-
-
-def _ds_read_tr8_b64_imm(result_type, addr_i32, imm_offset=0):
-    """gfx950 ds_read_b64_tr_b8 (8-bit transpose) with immediate byte offset.
-
-    Returns 64 bits = 8 fp8 (the fp8 analog of ds_read_b64_tr_b16's 4 bf16),
-    used for the fp8 V transpose load.
-
-    LOAD-BEARING asm: ``fx.rocdl.cdna4.LDSReadTrans8_64b()`` is the equivalent
-    copy atom, but this flydsl exposes no immediate-offset state for it, so
-    ``fx.copy`` over ``fx.add_offset`` views materialises one GEP per read
-    (54 vs 20 in the kernel IR) instead of folding into the instruction's
-    ``offset:`` field. Measured cost of the atom form: +31% at D=192/Dv=192
-    (24 reads per V load), -2% at Dv=128. Revisit if the atom gains imm_offset.
-    """
-    imm = int(imm_offset)
-    raw_type = ir.VectorType.get([2], ir.IntegerType.get_signless(32))
-    raw = llvm.inline_asm(
-        raw_type,
-        [as_mlir_value(addr_i32)],
-        f"ds_read_b64_tr_b8 $0, $1 offset:{imm}\n",
-        "=v,v,~{memory}",
-        has_side_effects=True,
-    )
-    return vector.BitCastOp(result_type, raw).result
-
-
-def _concat_vectors(lhs, rhs):
-    lhs_vec = Vec(lhs)
-    rhs_vec = Vec(rhs)
-    return lhs_vec.shuffle(
-        rhs_vec,
-        list(range(lhs_vec.numel)) + [lhs_vec.numel + i for i in range(rhs_vec.numel)],
-    )
-
-
-def _bitcast_i32(value):
-    return as_mlir_value(fx.Float32(value).bitcast(fx.Int32).ir_value())
-
-
-def _bitcast_f32(value):
-    return as_mlir_value(fx.Int32(value).bitcast(fx.Float32).ir_value())
-
-
-def _attn_mask_vec2_imm(rel_i32, neg_inf_i32, thr_x, thr_y, x_ref_i32, y_ref_i32):
-    """Causal pair mask: ``rel < thr ? -inf : score``, on the f32 bit patterns.
-
-    Two independent selects sharing one ``rel`` operand. This replaced a
-    hand-written ``2x v_cmp_lt_i32 + 2x v_cndmask_b32`` asm block whose
-    ``=s,=s`` + tied ``2,3`` constraints pinned the masks in SGPRs and let each
-    cndmask reuse its source register. The fx form is an exact 1:1 lowering
-    (32 asm blocks -> 64 arith.cmpi + 64 arith.select in the kernel IR) and is
-    bit-identical; it only hands register allocation back to the compiler.
-
-    Measured trade vs the asm form: varlen cross-causal -3.56% (299.7 -> 289.0us,
-    the compiler schedules better once the instructions are not pinned), varlen
-    self-causal +0.94% (42.16 -> 42.56us, dense-diagonal masking is where the
-    saved v_mov paid off). Net positive and it removes an inline asm that had a
-    perfectly good fx equivalent, so the fx form is what ships.
-    """
-    rel = fx.Int32(rel_i32)
-    neg_inf = fx.Int32(neg_inf_i32)
-    out_x = (rel < fx.Int32(thr_x)).select(neg_inf, fx.Int32(x_ref_i32))
-    out_y = (rel < fx.Int32(thr_y)).select(neg_inf, fx.Int32(y_ref_i32))
-    return out_x.ir_value(), out_y.ir_value()
-
-
-def _reduction_pair(v_f32):
-    v_i32 = _bitcast_i32(v_f32)
-    pair_ty = ir.Type.parse("!llvm.struct<(i32, i32)>")
-    swapped = rocdl.permlane32_swap(pair_ty, v_i32, v_i32, False, True)
-    lhs_i32 = llvm.extractvalue(T.i32, swapped, [0])
-    rhs_i32 = llvm.extractvalue(T.i32, swapped, [1])
-    return _bitcast_f32(lhs_i32), _bitcast_f32(rhs_i32)
-
-
-def _anchor_scalar_f32(x):
-    """Pin a scalar f32 at the current source position (no-op asm)."""
-    x_ir = as_mlir_value(x)
-    return llvm.inline_asm(
-        x_ir.type,
-        [x_ir],
-        "",
-        "=v,0",
-        has_side_effects=True,
-    )
-
-
-def _anchor_v_o(traits, v_o):
-    """Pin v_o accumulators at the current source position."""
-    acc_irs = [as_mlir_value(v_o[dc]) for dc in range_constexpr(traits.D_CHUNKS)]
-    ret_ty = ir.Type.parse(
-        f"!llvm.struct<({', '.join(['vector<16xf32>'] * traits.D_CHUNKS)})>"
-    )
-    constraints = ",".join(
-        ["=v"] * traits.D_CHUNKS + [str(i) for i in range(traits.D_CHUNKS)]
-    )
-    ret = llvm.inline_asm(
-        ret_ty,
-        acc_irs,
-        "",
-        constraints,
-        has_side_effects=True,
-    )
-    return [
-        llvm.extractvalue(acc_irs[dc].type, ret, [dc])
-        for dc in range_constexpr(traits.D_CHUNKS)
-    ]
-
-
-def _anchor_v_p(traits, v_p, elem_dtype):
-    p_lo, p_hi = v_p
-    p_lo_all = _concat_vectors(p_lo[0], p_lo[1])
-    p_hi_all = _concat_vectors(p_hi[0], p_hi[1])
-    p_all = _concat_vectors(p_lo_all, p_hi_all)
-    p_all_ir = as_mlir_value(p_all)
-    p_all_anchored = llvm.inline_asm(
-        p_all_ir.type,
-        [p_all_ir],
-        "",
-        "=v,0",
-        has_side_effects=True,
-    )
-    p_vec = Vec(p_all_anchored, (traits.PV_K_STEPS * 2 * 8,), elem_dtype)
-    anchored_lo = []
-    anchored_hi = []
-    for pks in range_constexpr(traits.PV_K_STEPS):
-        lo_base = pks * 8
-        hi_base = traits.PV_K_STEPS * 8 + pks * 8
-        anchored_lo.append(
-            p_vec.shuffle(p_vec, [lo_base + i for i in range(8)]).ir_value()
-        )
-        anchored_hi.append(
-            p_vec.shuffle(p_vec, [hi_base + i for i in range(8)]).ir_value()
-        )
-    return anchored_lo, anchored_hi
-
-
-def _v_pair_to_vec32(v):
-    return _concat_vectors(v[0], v[1]).ir_value()
-
-
-def _v_vec32_to_pair(v):
-    v_vec = Vec(v, (32,), fx.Float32)
-    v_lo = v_vec.shuffle(v_vec, [i for i in range(16)]).ir_value()
-    v_hi = v_vec.shuffle(v_vec, [16 + i for i in range(16)]).ir_value()
-    return v_lo, v_hi
-
-
-def _v_p_to_vec32(v_p):
-    p_lo, p_hi = v_p
-    p_lo_all = _concat_vectors(p_lo[0], p_lo[1])
-    p_hi_all = _concat_vectors(p_hi[0], p_hi[1])
-    return _concat_vectors(p_lo_all, p_hi_all).ir_value()
-
-
-def _v_vec32_to_p(traits, v_p_all, elem_dtype):
-    p_vec = Vec(v_p_all, (traits.PV_K_STEPS * 2 * 8,), elem_dtype)
-    p_lo = []
-    p_hi = []
-    for pks in range_constexpr(traits.PV_K_STEPS):
-        lo_base = pks * 8
-        hi_base = traits.PV_K_STEPS * 8 + pks * 8
-        p_lo.append(p_vec.shuffle(p_vec, [lo_base + i for i in range(8)]).ir_value())
-        p_hi.append(p_vec.shuffle(p_vec, [hi_base + i for i in range(8)]).ir_value())
-    return p_lo, p_hi
-
-
-def _score_pair_to_lists(v_s):
-    s_lo, s_hi = v_s
-    return (
-        [Vec(s_lo)[r] for r in range_constexpr(16)],
-        [Vec(s_hi)[r] for r in range_constexpr(16)],
-    )
-
-
-def _score_lists_to_vecs(v_s_lists):
-    s_lo, s_hi = v_s_lists
-    return (
-        Vec.from_elements([as_mlir_value(v) for v in s_lo], fx.Float32).ir_value(),
-        Vec.from_elements([as_mlir_value(v) for v in s_hi], fx.Float32).ir_value(),
-    )
-
-
-def _reduce_score_pair(v_s, initial, reducer, fm_fast):
-    s_lo, s_hi = v_s
-    acc = initial
-    for r in range_constexpr(16):
-        acc = reducer(acc, s_lo[r], fm_fast)
-    for r in range_constexpr(16):
-        acc = reducer(acc, s_hi[r], fm_fast)
-    return acc
-
-
-def _lane_pair_reduce(v, reducer, fm_fast):
-    lhs, rhs = _reduction_pair(v)
-    return reducer(lhs, rhs, fm_fast)
-
-
-def _score_pair_max(v_s, neg_inf, fm_fast):
-    reducer = lambda a, b, _fm: fx.maxnumf(a, b)
-    return _lane_pair_reduce(
-        _reduce_score_pair(v_s, neg_inf, reducer, fm_fast), reducer, fm_fast
-    )
-
-
-def _score_pair_sum(v_s, zero_f, fm_fast):
-    s_lo, s_hi = _score_lists_to_vecs(v_s)
-    tile = Vec(s_lo) + Vec(s_hi)
-    reducer = lambda a, b, _fm: a + b
-    return _lane_pair_reduce(
-        tile.reduce("add", init_val=zero_f, fastmath=fm_fast), reducer, fm_fast
-    )
-
-
-def _scale_sub_score_pair(v_s, row_max_raw, scale, zero_f, fm_fast, bias=None):
-    """Fused softmax-scale + row-max subtraction (optimization 1-A).
-
-    Returns ``scale * (v_s - row_max_raw) + bias`` per element via a single FMA
-    (``fma(s, scale, bias - scale*row_max_raw)``), so the fp8 QK MMA can emit raw
-    (un-scaled) logits and reduce_max can run in the raw domain (scale > 0 is
-    order-preserving). Replaces the separate post-QK scale multiply + subtract.
-    ``-inf`` masked lanes stay ``-inf`` (scale > 0), matching the un-fused path.
-
-    ``bias`` lands in the FMA's addend, so a caller needing ``exp2`` to produce
-    ``2**bias * P`` pays nothing -- see ``DualwaveFp8SoftmaxHelper.sub_m``.
-    """
-    s_lo, s_hi = v_s
-    neg_scaled_max = zero_f - scale * row_max_raw
-    if bias is not None:
-        # exp2 lands on 2**bias * P instead of P, at no extra instruction: the
-        # FMA's addend absorbs it.
-        neg_scaled_max = neg_scaled_max + bias
-    scale_v = Vec.from_elements([scale], fx.Float32).broadcast_to(16)
-    nsm_v = Vec.from_elements([neg_scaled_max], fx.Float32).broadcast_to(16)
-    lo = fx.fma(Vec(s_lo), scale_v, nsm_v, fastmath=fm_fast)
-    hi = fx.fma(Vec(s_hi), scale_v, nsm_v, fastmath=fm_fast)
-    return as_mlir_value(lo), as_mlir_value(hi)
-
-
-def _exp2_score_slice(v_s, start):
-    if const_expr(start == 0):
-        s_lo = [Vec(v_s[0])[r] for r in range_constexpr(16)]
-        lo_partial = []
-        for r in range_constexpr(16):
-            lo_partial.append(rocdl.exp2(T.f32, as_mlir_value(s_lo[r])))
-        return Vec.from_elements(lo_partial, fx.Float32).ir_value(), v_s[1]
-
-    lo_partial = [Vec(v_s[0])[r] for r in range_constexpr(16)]
-    hi_full = []
-    for r in range_constexpr(16):
-        hi_full.append(rocdl.exp2(T.f32, as_mlir_value(Vec(v_s[1])[r])))
-    return lo_partial, hi_full
-
-
-def _pack_p_v8_slices(traits, v_p, pack_v8_fn):
-    lo_partial_list, hi_full = v_p
-    p_lo_packs = []
-    p_hi_packs = []
-    for pks in range_constexpr(traits.PV_K_STEPS):
-        p_base = pks * 8
-        lo_slice = [lo_partial_list[p_base + s] for s in range_constexpr(8)]
-        hi_slice = hi_full[p_base : p_base + 8]
-        p_lo_packs.append(pack_v8_fn(lo_slice))
-        p_hi_packs.append(pack_v8_fn(hi_slice))
-    return p_lo_packs, p_hi_packs
-
-
-def _safe_l_inv(l_row, zero_f):
-    l_inv = rocdl.rcp(T.f32, as_mlir_value(l_row))
-    return (fx.Float32(l_row) > zero_f).select(l_inv, zero_f)
-
-
-def _scale_o_accs(v_o, scale_scalar, traits):
-    scale_vec = Vec.from_elements([scale_scalar], fx.Float32).broadcast_to(16)
-    for dc in range_constexpr(traits.D_CHUNKS):
-        v_o[dc] = Vec(v_o[dc]) * scale_vec
-
-
-def _causal_pair_thresholds(kv_vectorized):
-    if const_expr(kv_vectorized):
-        return [
-            (0, 1),
-            (2, 3),
-            (4, 5),
-            (6, 7),
-            (16, 17),
-            (18, 19),
-            (20, 21),
-            (22, 23),
-        ]
-    return [
-        (0, 1),
-        (2, 3),
-        (8, 9),
-        (10, 11),
-        (16, 17),
-        (18, 19),
-        (24, 25),
-        (26, 27),
-    ]
-
-
-def _apply_dualwave_causal_mask_pair(s_values, rel_i32, neg_inf_i32, pair_thresholds):
-    for p in range_constexpr(len(pair_thresholds)):
-        thr_x, thr_y = pair_thresholds[p]
-        idx_x = p * 2
-        idx_y = p * 2 + 1
-        x_bits = _bitcast_i32(s_values[idx_x])
-        y_bits = _bitcast_i32(s_values[idx_y])
-        new_x, new_y = _attn_mask_vec2_imm(
-            rel_i32, neg_inf_i32, thr_x, thr_y, x_bits, y_bits
-        )
-        s_values[idx_x] = _bitcast_f32(new_x)
-        s_values[idx_y] = _bitcast_f32(new_y)
-
-
-def _cu_load(div, idx, cu_atom, cu_v1i32):
-    """Load cu_seqlens[idx] into an SGPR. ``idx`` must be wave-uniform."""
-    v = fly.copy_atom_call_ssa(
-        [cu_v1i32], cu_atom, fx.slice(div, (None, fx.Int32(idx)))
-    )
-    return fx.Index(
-        rocdl.readfirstlane(T.i32, as_mlir_value(fx.Int32(Vec(v, (1,), fx.Int32)[0])))
-    )
-
-
-def _make_ws_rsrc(ws_base_i64, byte_offset, nrec_bytes):
-    addr_i64 = as_mlir_value(ws_base_i64 + fx.Int64(byte_offset))
-    return buffer_ops.create_buffer_resource_from_addr(
-        addr_i64, num_records_bytes=as_mlir_value(fx.Int64(nrec_bytes))
-    )
-
-
-def _buffer_load_128(elem_index, _load_atom_128, q_div, q_load_i32x4_type):
-    """128-bit global->register load (buffer_load_dwordx4) from Q."""
-    return fly.copy_atom_call_ssa(
-        [q_load_i32x4_type],
-        _load_atom_128,
-        fx.slice(q_div, (None, fx.Int32(elem_index))),
-    )
-
-
-def _buffer_load_lds_128(
-    src_div, lds_byte_addr, src_elem, soffset_elems, _dma_atom, _lds_ptr_ty
-):
-    """128-bit global->LDS DMA; `src_elem` is voffset, `soffset_elems` is scaled by the atom."""
-    lds_ptr = fx.inttoptr(_lds_ptr_ty, fx.Int32(lds_byte_addr))
-    dst = fx.make_view(lds_ptr, fx.make_layout(1, 1))
-    src = fx.slice(src_div, (None, fx.Int32(src_elem)))
-    fx.copy(_dma_atom, src, dst, soffset=fx.Int32(soffset_elems))
-
-
-def _buffer_store_128(
-    pack_i32_vec, elem_index, _o_store_reg_128, _store_atom_128, o_div
-):
-    """128-bit register->global store (buffer_store_dwordx4) into O."""
-    fx.memref_store_vec(pack_i32_vec, _o_store_reg_128)
-    fx.copy(
-        _store_atom_128, _o_store_reg_128, fx.slice(o_div, (None, fx.Int32(elem_index)))
-    )
-
-
-@flyc.jit
-def _stagger_extra_barrier_if_one(stagger_i32):
-    """Emit `sched_barrier(0); s_barrier;` only when stagger == 1."""
-    if fx.Int32(stagger_i32) != fx.Int32(0):
-        rocdl.sched_barrier(0)
-        rocdl.s_barrier()
