+++ aiter/ops/flydsl/kernels/fmha_gfx950/op_combine.py
@@ -2,12 +2,12 @@
 # Copyright (c) 2025 FlyDSL Project Contributors
 # Modifications Copyright (C) 2026 Advanced Micro Devices, Inc.
 
-"""Split-K workspace sizing and the combine pass.
+"""Split-K COMBINE pass and its workspace sizing.
 
 Part of the gfx950 dual-wave fp8 (e4m3fn) flash-attention kernel, migrated from
 FlyDSL ``kernels/attention/flash_attn_utils.py`` and restricted to the symbols
-the fp8 path reaches. The bf16/f16 dual-wave, the gfx942 generic path, paged KV,
-and the bias/ALiBi helpers are not part of it and were left behind.
+the fp8 path reaches. File layout follows csrc/kernels/mha_native/fused:
+``op_*`` per stage, ``pipeline`` for the per-block forward pass.
 """
 
 import flydsl.compiler as flyc
@@ -19,7 +19,15 @@ from flydsl.expr.typing import Vector as Vec
 from flydsl.expr.utils.arith import _to_raw as as_mlir_value
 
 from aiter.ops.flydsl.kernels import buffer_ops
-from aiter.ops.flydsl.kernels.fmha_gfx950.flash_attn_utils_primitives import (
+
+"""Split-K workspace sizing and the combine pass.
+
+Part of the gfx950 dual-wave fp8 (e4m3fn) flash-attention kernel, migrated from
+FlyDSL ``kernels/attention/flash_attn_utils.py`` and restricted to the symbols
+the fp8 path reaches. The bf16/f16 dual-wave, the gfx942 generic path, paged KV,
+and the bias/ALiBi helpers are not part of it and were left behind.
+"""
+from aiter.ops.flydsl.kernels.fmha_gfx950.pipeline import (
     _LOG2E,
     _cu_load,
     _make_ws_rsrc,
