+++ aiter/ops/flydsl/kernels/fmha_gfx950/op_epilog.py
@@ -1,0 +1,147 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2025 FlyDSL Project Contributors
+# Modifications Copyright (C) 2026 Advanced Micro Devices, Inc.
+
+"""Epilogue: output packing and the dense / split-K partial stores.
+
+Part of the gfx950 dual-wave fp8 (e4m3fn) flash-attention kernel, migrated from
+FlyDSL ``kernels/attention/flash_attn_utils.py`` and restricted to the symbols
+the fp8 path reaches. File layout follows csrc/kernels/mha_native/fused:
+``op_*`` per stage, ``pipeline`` for the per-block forward pass.
+"""
+
+import flydsl.compiler as flyc
+import flydsl.expr as fx
+from flydsl._mlir import ir
+from flydsl._mlir.dialects import llvm
+from flydsl.expr import gpu, range_constexpr, rocdl
+from flydsl.expr.typing import T
+from flydsl.expr.typing import Vector as Vec
+from flydsl.expr.utils.arith import _to_raw as as_mlir_value
+
+from aiter.ops.flydsl.kernels.fmha_gfx950.pipeline import DualwaveFp8KernelContext
+
+
+class DualwaveFp8StoreHelper(DualwaveFp8KernelContext):
+    def __init__(self, ctx):
+        super().__init__(ctx)
+
+    def _o_pack_2dw(self, v_o, dc, store_group):
+        r_base = store_group * 4
+        lo = rocdl.cvt_pk_bf16_f32(Vec(v_o[dc])[r_base], Vec(v_o[dc])[r_base + 1])
+        hi = rocdl.cvt_pk_bf16_f32(Vec(v_o[dc])[r_base + 2], Vec(v_o[dc])[r_base + 3])
+        return lo, hi
+
+    def _swap_half_partner(self, dw):
+        pair_i32_ty = ir.Type.parse("!llvm.struct<(i32, i32)>")
+        swapped = rocdl.permlane32_swap(
+            pair_i32_ty, as_mlir_value(dw), as_mlir_value(dw), False, False
+        )
+        lo_res = llvm.extractvalue(T.i32, swapped, [0])
+        hi_res = llvm.extractvalue(T.i32, swapped, [1])
+        return (self.lane_div_32 != 0).select(lo_res, hi_res)
+
+    def _packed_o_128_dwords(self, v_o, dc, g):
+        is_hi_half = self.lane_div_32 != 0
+        d0_a, d1_a = self._o_pack_2dw(v_o, dc, 2 * g)
+        d0_b, d1_b = self._o_pack_2dw(v_o, dc, 2 * g + 1)
+        y0_a, y1_a = self._swap_half_partner(d0_a), self._swap_half_partner(d1_a)
+        y0_b, y1_b = self._swap_half_partner(d0_b), self._swap_half_partner(d1_b)
+        w0 = is_hi_half.select(y0_b, as_mlir_value(d0_a))
+        w1 = is_hi_half.select(y1_b, as_mlir_value(d1_a))
+        w2 = is_hi_half.select(as_mlir_value(d0_b), y0_a)
+        w3 = is_hi_half.select(as_mlir_value(d1_b), y1_a)
+        return w0, w1, w2, w3
+
+    def _packed_o_128_vec(self, v_o, dc, g):
+        return Vec.from_elements(
+            [fx.Int32(w) for w in self._packed_o_128_dwords(v_o, dc, g)], fx.Int32
+        )
+
+    def store_final_o(self, v_o, q_row):
+        for dc in range_constexpr(self.traits.D_CHUNKS):
+            for g in range_constexpr(2):
+                o_pack = self._packed_o_128_vec(v_o, dc, g)
+                d_col = (dc * self.traits.D_CHUNK) + (2 * g + self.lane_div_32) * 8
+                o_global = self.global_idx_o(q_row, d_col)
+                self.buffer_store_128(o_pack, o_global)
+
+    def store_splitk_partial_o(self, v_o, m_row, l_row, q_row):
+        m_row = fx.Float32(m_row) * self.c_logit_scale
+        split_z = self.batch_idx * self.traits.NUM_KV_SPLITS + self.split_idx
+        o_part_row_base = (
+            (split_z * self.traits.NUM_HEADS_Q + self.q_head_idx) * self.seq_len_v
+            + q_row
+        ) * (self.traits.HEAD_DIM_V // 2)
+        grid_z = fx.Index(gpu.grid_dim.z)
+        mrow_base = (
+            grid_z
+            * self.traits.NUM_HEADS_Q
+            * self.seq_len_v
+            * (self.traits.HEAD_DIM_V // 2)
+        )
+        lrow_base = mrow_base + grid_z * self.traits.NUM_HEADS_Q * self.seq_len_v
+        ml_row_idx = (
+            split_z * self.traits.NUM_HEADS_Q + self.q_head_idx
+        ) * self.seq_len_v + q_row
+
+        @flyc.jit
+        def _store_splitk_partial_if_qrow():
+            if q_row < self.seq_len_v:
+                for dc in range_constexpr(self.traits.D_CHUNKS):
+                    for g in range_constexpr(2):
+                        dw_col = (
+                            dc * (self.traits.D_CHUNK // 2)
+                            + (2 * g + self.lane_div_32) * 4
+                        )
+                        self.ws_store_quad_i32(
+                            self._packed_o_128_dwords(v_o, dc, g),
+                            o_part_row_base + dw_col,
+                        )
+                if self.lane < 32:
+                    self.ws_store_f32(m_row, mrow_base + ml_row_idx)
+                    self.ws_store_f32(l_row, lrow_base + ml_row_idx)
+
+        _store_splitk_partial_if_qrow()
+
+    def store_empty_split(self):
+        @flyc.jit
+        def _store_empty_split():
+            if self.max_num_tiles < self.split_t0 + 4:
+                q_row_e = self.q_start + self.wave_q_offset + self.lane_mod_32
+                split_z_e = self.batch_idx * self.traits.NUM_KV_SPLITS + self.split_idx
+                o_row_base_e = (
+                    (split_z_e * self.traits.NUM_HEADS_Q + self.q_head_idx)
+                    * self.seq_len_v
+                    + q_row_e
+                ) * (self.traits.HEAD_DIM_V // 2)
+                grid_z_e = fx.Index(gpu.grid_dim.z)
+                mrow_base_e = (
+                    grid_z_e
+                    * self.traits.NUM_HEADS_Q
+                    * self.seq_len_v
+                    * (self.traits.HEAD_DIM_V // 2)
+                )
+                lrow_base_e = (
+                    mrow_base_e + grid_z_e * self.traits.NUM_HEADS_Q * self.seq_len_v
+                )
+                ml_row_e = (
+                    split_z_e * self.traits.NUM_HEADS_Q + self.q_head_idx
+                ) * self.seq_len_v + q_row_e
+                if q_row_e < self.seq_len_v:
+                    c_zero_i = fx.Int32(0)
+                    for dc in range_constexpr(self.traits.D_CHUNKS):
+                        for g in range_constexpr(2):
+                            dw_col = (
+                                dc * (self.traits.D_CHUNK // 2)
+                                + (2 * g + self.lane_div_32) * 4
+                            )
+                            self.ws_store_quad_i32(
+                                [c_zero_i, c_zero_i, c_zero_i, c_zero_i],
+                                o_row_base_e + dw_col,
+                            )
+                    if self.lane < 32:
+                        self.ws_store_f32(fx.Float32(-1e30), mrow_base_e + ml_row_e)
+                        self.ws_store_f32(self.c_zero_f, lrow_base_e + ml_row_e)
+
+        _store_empty_split()
