+++ aiter/ops/flydsl/kernels/fmha_gfx950/op_lds.py
@@ -1,0 +1,221 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2025 FlyDSL Project Contributors
+# Modifications Copyright (C) 2026 Advanced Micro Devices, Inc.
+
+"""LDS (shared-memory) data movement: Q load, K/V global->LDS, LDS->VGPR.
+
+Part of the gfx950 dual-wave fp8 (e4m3fn) flash-attention kernel, migrated from
+FlyDSL ``kernels/attention/flash_attn_utils.py`` and restricted to the symbols
+the fp8 path reaches. File layout follows csrc/kernels/mha_native/fused:
+``op_*`` per stage, ``pipeline`` for the per-block forward pass.
+"""
+
+import flydsl.compiler as flyc
+import flydsl.expr as fx
+from flydsl._mlir.dialects import llvm
+from flydsl.expr import const_expr, range_constexpr
+from flydsl.expr.typing import T
+from flydsl.expr.typing import Vector as Vec
+from flydsl.expr.utils.arith import _to_raw as as_mlir_value
+
+from aiter.ops.flydsl.kernels.fmha_gfx950.pipeline import (
+    DualwaveFp8KernelContext,
+    _ds_read_tr8_b64_imm,
+)
+
+
+class DualwaveFp8QLoader(DualwaveFp8KernelContext):
+    def __init__(self, ctx):
+        super().__init__(ctx)
+
+    def stage_q_to_lds(self):
+        traits = self.traits
+        chunks_per_row = traits.HEAD_DIM // 16  # 16-byte DMA chunks per Q row
+        total_chunks = (traits.BLOCK_M * traits.HEAD_DIM) // 16
+        for p in range_constexpr(total_chunks // traits.BLOCK_SIZE):
+            c = self.tid + (p * traits.BLOCK_SIZE)
+            row = c // chunks_per_row
+            dchunk = c % chunks_per_row
+            src_elem = self.q_gmem_elem_offset + row * self.stride_q_n_v + dchunk * 16
+            lds_addr = self.lds_q_base_idx + c * 16
+            self.buffer_load_lds_128(self.q_div, lds_addr, src_elem, 0)
+
+    def load_all_wide(self, q_row_in_block):
+        traits = self.traits
+        d_base = self.lane_div_32 * 32
+        packs = []
+        for ws in range_constexpr(traits.HEAD_DIM // 64):
+            byte_row = q_row_in_block * traits.HEAD_DIM + (ws * 64) + d_base
+            packs.append(self.read_i32x8_lds(self.lds_q_base_ptr, fx.Int32(byte_row)))
+        return packs
+
+
+class DualwaveFp8KvGmemToLdsLoader(DualwaveFp8KernelContext):
+    def __init__(self, ctx):
+        super().__init__(ctx)
+
+    def load_k(self, tile_start, buf_id):
+        """DMA one K tile into LDS, one pass per head-dim band.
+
+        A band's LDS line is this wave's n-rows of `chunk` bytes, row-contiguous,
+        so the QK read indexes it as (band, row, 64-byte slice). The 64-byte tail
+        band at head_dim 192 runs on the low 32 lanes and moves no padding.
+        """
+        traits = self.traits
+        eb = traits.ELEM_BYTES
+        k_lds_byte_base = self.lds_kv_base_idx + self.k_buf_base(buf_id) * eb
+        rows_per_wave = -(-traits.BLOCK_N // traits.NUM_WAVES)
+        for d in range_constexpr(self.NUM_DMA_K):
+            lanes_per_row = traits.K_BAND_CHUNK[d] // traits.VEC_KV
+            slots = rows_per_wave * lanes_per_row
+            band_base = (
+                k_lds_byte_base
+                + traits.K_BAND_BASE[d] * eb
+                + self.wave_id_uni * (traits.K_BAND_LINE_STRIDE[d] * eb)
+            )
+            for pas in range_constexpr(-(-slots // traits.WARP_SIZE)):
+                slot = self.lane_in_warp + (pas * traits.WARP_SIZE)
+                n_in_tile = (slot // lanes_per_row) * traits.NUM_WAVES + self.wave_id
+                global_d = (
+                    slot % lanes_per_row
+                ) * traits.VEC_KV + traits.K_BAND_GLOBAL_D[d]
+                src_elem = (
+                    self.kv_gmem_elem_offset + n_in_tile * self.stride_kv_n_v + global_d
+                )
+                lds_addr = band_base + fx.Index(
+                    pas * traits.WARP_SIZE * traits.VEC_KV * eb
+                )
+                active = min(slots - pas * traits.WARP_SIZE, traits.WARP_SIZE)
+                if const_expr(active == traits.WARP_SIZE):
+                    self.buffer_load_lds_128(
+                        self.k_div, lds_addr, src_elem, tile_start * self.stride_kv_n_v
+                    )
+                else:
+                    self._load_k_band_partial_wave(
+                        lds_addr, src_elem, tile_start, active
+                    )
+
+    def _load_k_band_partial_wave(self, lds_addr, src_elem, tile_start, active_lanes):
+        soffset = tile_start * self.stride_kv_n_v
+        k_div = self.k_div
+
+        @flyc.jit
+        def _run():
+            if self.lane_in_warp < active_lanes:
+                self.buffer_load_lds_128(k_div, lds_addr, src_elem, soffset)
+
+        _run()
+
+    def load_v(self, tile_start, buf_id):
+        self._stage_v_fp8_block_dma(tile_start, buf_id)
+
+    def _stage_v_fp8_block_dma(self, tile_start, buf_id):
+        traits = self.traits
+        nbands = traits.HEAD_DIM_V // 16
+        v_tile_bytes = (traits.BLOCK_N // 8) * nbands * 128
+        buf_off = buf_id * v_tile_bytes
+        aligned_base = ((self.lds_vt_base_idx + 127) // 128) * 128
+        # The tile is BLOCK_N * nbands 16-byte slots, and one DMA instruction moves a
+        # whole wave of them. Hand out instructions, not row-groups: a wave's LDS
+        # destination is then always a full WARP_SIZE*16 span, so nothing has to be
+        # masked off inside a wave. buffer_load...lds strides the LDS write by lane
+        # regardless of exec, so an intra-wave mask would still write past the span.
+        per_dma = traits.WARP_SIZE * traits.VEC_KV * traits.ELEM_BYTES
+        slots_per_group = 8 * nbands
+        num_dma = (traits.BLOCK_N * nbands * 16) // per_dma
+        passes = -(-num_dma // traits.NUM_WAVES)
+        for pas in range_constexpr(passes):
+            dma_id = self.wave_id_uni + (pas * traits.NUM_WAVES)
+            slot = dma_id * traits.WARP_SIZE + self.lane
+            lds_addr = aligned_base + fx.Index(buf_off) + dma_id * per_dma
+            grp = slot // slots_per_group
+            rem = slot % slots_per_group
+            dest_n = fx.Int32(grp * 8 + rem % 8)
+            w16 = dest_n % fx.Int32(16)
+            c_add = (w16 >= fx.Int32(4)) & (w16 < fx.Int32(8))
+            c_sub = (w16 >= fx.Int32(8)) & (w16 < fx.Int32(12))
+            n = (
+                dest_n
+                + c_add.select(fx.Int32(4), fx.Int32(0))
+                - c_sub.select(fx.Int32(4), fx.Int32(0))
+            )
+            d_block = rem // 8
+            src_elem = (
+                self.v_gmem_elem_offset + fx.Index(n) * self.stride_v_n_v + d_block * 16
+            )
+            if const_expr(num_dma % traits.NUM_WAVES == 0 or pas < passes - 1):
+                self.buffer_load_lds_128(
+                    self.v_div, lds_addr, src_elem, tile_start * self.stride_v_n_v
+                )
+            else:
+                self._load_v_group_if_in_tile(
+                    lds_addr, src_elem, tile_start, dma_id, num_dma
+                )
+
+    def _load_v_group_if_in_tile(self, lds_addr, src_elem, tile_start, grp, groups):
+        soffset = tile_start * self.stride_v_n_v
+        v_div = self.v_div
+
+        @flyc.jit
+        def _run():
+            if grp < groups:
+                self.buffer_load_lds_128(v_div, lds_addr, src_elem, soffset)
+
+        _run()
+
+
+class DualwaveFp8KvLdsToVgprLoader(DualwaveFp8KernelContext):
+    def __init__(self, ctx):
+        super().__init__(ctx)
+
+    def load_k(self, buf_id):
+        # Read K in the wide 32x32x64 QK operand layout (32 contiguous head-dim/lane,
+        # two N-strips, two head-dim halves).
+        traits = self.traits
+        k_base = self.k_buf_base(buf_id)
+        d_base = self.lane_div_32 * 32
+        n_lo = self.lane_mod_32
+        n_hi = self.lane_mod_32 + 32
+
+        rows_per_line = traits.NUM_WAVES
+
+        def _read_strip(key):
+            out = []
+            for ws in range_constexpr(traits.HEAD_DIM // 64):
+                b = traits.K_WS_BAND[ws]
+                line = (key % rows_per_line) * traits.K_BAND_LINE_STRIDE[b]
+                row = line + (key // rows_per_line) * traits.K_BAND_CHUNK[b]
+                addr = (
+                    k_base + traits.K_BAND_BASE[b] + row + traits.K_WS_OFF[ws] + d_base
+                )
+                out.append(self.read_i32x8_lds(self.lds_kv_base_ptr, addr))
+            return out
+
+        return (_read_strip(n_lo), _read_strip(n_hi))
+
+    def load_v(self, buf_id):
+        return self._load_v_fp8_block(buf_id)
+
+    def _load_v_fp8_block(self, buf_id):
+        traits = self.traits
+        v_tile_bytes = (traits.BLOCK_N // 8) * (traits.HEAD_DIM_V // 16) * 128
+        buf_off = buf_id * v_tile_bytes
+        nbands = traits.HEAD_DIM_V // 16
+        rh = (self.lane % 32) // 16
+        l16 = self.lane % 16
+        lane_hi = self.lane // 32
+        aligned_base = ((self.lds_vt_base_idx + 127) // 128) * 128
+        base = fx.Int32(
+            aligned_base + buf_off + rh * 128 + l16 * 8 + lane_hi * (nbands * 128)
+        )
+
+        def _tr8(imm):
+            r = _ds_read_tr8_b64_imm(self.v2i32_type, base, imm)
+            return llvm.bitcast(T.i64, as_mlir_value(Vec(r)))
+
+        packs = [[None] * traits.D_CHUNKS for _ in range(4)]
+        for dc in range_constexpr(traits.D_CHUNKS):
+            for ks in range_constexpr(4):
+                imm0 = (2 * ks * nbands + dc * 2) * 128
+                packs[ks][dc] = _tr8(imm0)
+        return packs
